Remove `RevocationBitmapService`
