+++ identity-did/src/service/mod.rs
@@ -4,11 +4,9 @@
 #![allow(clippy::module_inception)]
 
 mod builder;
-mod revocation_bitmap_service;
 mod service;
 mod service_endpoint;
 
 pub use self::builder::ServiceBuilder;
-pub use self::revocation_bitmap_service::RevocationBitmapService;
 pub use self::service::Service;
 pub use self::service_endpoint::ServiceEndpoint;
